Bug 576587 – allow eject even on non-ejectable volumes
This fixes a host of problems with e.g. Kindle or iPod devices
requiring to be "ejected" to display messages such as

 "If you want to use your Kindle and continue charging, please eject
 your Kindle from your computer."

to the user.

Previously we relied on HAL fdi files or similar to tag that such
devices needed to be ejectable. Now we just set everything as
ejectable.

For this to really work well the Nautilus patch in

 http://bugzilla.gnome.org/show_bug.cgi?id=574067

needs to be reverted.