Pass the 'flush' mount option for vfat
This mount option makes the vfat filesystem driver flush data more
often. As a consequence

 1. users never get to see the gnome-disk-utility notification daemon
    dialog just added

     http://people.freedesktop.org/~david/gdu-unmount-busy-1.png

    but that's useful for other filesystems as well.

 2. The Nautilus copy dialog stays up until things are on the disk

We do this in gvfs rather than DeviceKit-disks because in some
scenarios 'flush' may be unwanted and there is currently no way to
turn it off (e.g. no 'noflush' or 'flush=0' option).

Ideally the kernel would get this kind of thing right by itself.