[FTP] split out connection code
the current FtpConnection object is not flexible enough. It is doing too
many things at once, namely:
1) handling a physical connection to the ftp server
2) handling a GVfsJob's lifetime
Also, the file gvfsbackendftp.c is too big, counting >2000 lines.

This is the first step in an attempt to separate these two things. It
splits out the physical connection into different files and makes the
old FtpConnection code use this instead.
A side effect of this is that it includes a change so that reading lines
is now done with a GDataInputStream instead of manually, as suggested by
Alex.