[FTP] keep list of busy connections
A busy connection is a connection that is currently used as a handle for
a file transfer (upload or download). If all connections are busy, we
can just return EBUSY instead of waiting for 30 seconds for a connection
to become available.

This is mostly useful for performance when there is a maximum of 1
connection to the server and this connection is busy transferring a
file. Even gio functions like g_file_copy() sometimes do additional
operations while transferring files. (It does a query_info for the
progress callback.)