gdu: Use DriveDetach() method to power down hard disk enclosures
See this commit

http://cgit.freedesktop.org/DeviceKit/DeviceKit-disks/commit/?id=5b7350df8b63cf4178a2fe69ee740b5222130421

for details about DriveDetach(). In a nutshell it allows us to offer
an Eject() option on USB enclosures. Currently we don't do this since
such devices don't have removable media.

With this change, we'll set can_eject to TRUE if the device is
detachable. If the device is both detachable and removable, we prefer
Eject over Detach() when doing the deed (e.g. consider a USB optical
drive; it is both detachable and removable).

To sum up, the net result of this change is that users of Nautilus
will get an Eject option and the eject icon in Nautilus' sidebar. Upon
the user pressing Eject in the UI we'll unmount all partitions as
usual. Then if the device is not removable, we'll invoke
DriveDetach(). That will make DeviceKit-disks send these two SCSI
commands

 SYNCHRONIZE CACHE
 STOP

and then we'll unbind e.g. the usb-storage driver from the USB
interface of the USB device. The net result is that the device appears
to be disconnected (ie. we get remove uevents for e.g. /dev/sdb[0..N]
and /dev/sdb) and also that the device powers down.

Right now DeviceKit-disks only supports DriveDetach() for USB devices
but should it gain support for doing the same thing for Firewire,
eSATA, whatever then we'll get that support for free too.