[FTP] rework handling of uncached files
The directory cache needs a callback to lookup uncachable files.
Uncachaable files are files that can be accessed but where the parent
directory cannot be listed.
This has to go into the directory cache, as the directory cache performs
symlink resolving and looking up symlinks may encounter such files.

Fixes a testcase in Andreas' test collection.