signal the need for updates (changes wire protocol and internal API)
Send a boolean send_progress to the daemon in the case of push/pull jobs
that indicates if progress updates should be sent to the client.
Oftentimes the daemons can avoid quite a bit of work (like querying file
sizes or setting up and operating machinery required to send progress
updates) when it's not required.

Patch also includes fixes to daemons to ensure they don't call a NULL
progress_callback (previously, NULL was not a possible value).