Avoid deadlock on cancelling async enumerate_children
We can't call g_cancellable_disconnect from inside the cancel signal
emission, as that deadlocks. However we know if that happens and can
use the regular disconnect.

Also, we need to grab the infos lock when calling trigger_async_done, this
was only done in some cases, now its always done.