port gphoto2 backend and monitor to gudev
This works with the ID_GPHOTO2 attribute, which was recently added to
libgphoto2:
http://sourceforge.net/tracker/?func=detail&aid=2801117&group_id=8874&atid=308874

That initial libgphoto2 patch does not yet pass the actual name, and just sets
ID_GPHOTO2="1". However, in this patch we check if it's not "1", and use that
as volume name (since it's the "best" one). Otherwise fall back to usb_id and
sysfs vendor/model/product names.

This also supports ID_MEDIA_PLAYER* attributes for media players with gphoto
support. Please see here for the discussion and development progress:
http://lists.freedesktop.org/archives/devkit-devel/2009-June/000226.html
http://cgit.freedesktop.org/~teuf/media-player-id/

http://bugzilla.gnome.org/show_bug.cgi?id=586410

Signed-off-by: David Zeuthen <davidz@redhat.com>