Avoid deadlock when pulling resolved dns-sd record from cache
When the host has already been resolved and is present in cache,
it's returned immediately. But we always started a mainloop
resulting in endless waiting for an event which had been received
before that. This applies for the sync call.

This is just a workaround, the sync code should be ported over avahi.
See bug 555436 (comments 30-32) for details.