Keep track of connection usability for new commands
Add a waiting_for_reply boolean flag to the connection. Set it when
sending a request to the server, clear it after receiving a reply.
When trying to push the connection into the connection pool, only allow
that when it is not waiting for a reply.
This way we fix the case where the task was cancelled while waiting for
the reply. It would then ignore the reply and happily push the connection
with the pending reply back into the pool, leaving the next one to use
it with a stale reply. You can imagine the hilarity that happened from
this.

Also, g_return_if_fail() if trying to send a command when waiting for a
reply or when trying to receive a command while not waiting.
If it turns out I overlooked a valid case where this can happen, we
need to remove those checks. So far I didn't find any.