[gdu] Be more careful with automounting - use a whitelist
The current behavior is that we try to automount all filesystems. This
has unintended consequences for big iron boxes connected to a SAN when
the user is logging in as root (thus having all polkit authorizations
at hand) - literally hundreds or thousands of devices may get mounted.

As such, only automount filesystems from the current whitelist

 - anything connected via USB or Firewire or SDIO buses
 - optical discs

Signed-off-by: David Zeuthen <davidz@redhat.com>