Attach cancellable to async result and check in finish functions
This makes sure we always return ERROR_CANCELLED, just like we want to
guarantee and avoids crashes in the file chooser.

There should be API in GSimpleAsyncResult to attach a GCancellable that
does this job for us, but so far there isn't, so we use
g_object_get/set_data() and check the cancellables manually in the
finish function.

https://bugzilla.gnome.org/show_bug.cgi?id=614099