dav: correctly handle usernames supplied in the url
When comparing URIs in dav_uri_match() do not take username or
password into account because on a redirect the server is likely to not
include the username in the url supplied in the Location header.
Therefore redirection will fail and thus mounting.
Should fix bug 586248. Patch based on a version from Henrik Holst.