dav: Include username and/or port in the backend's display-name
If there was a username given by the user and/or the  user specified a (non-default)
port we need to take those into account when creating the share's display-name. This
will resolve ambiguities that could otherwise happen if we have two mounts with
different users or on different ports to the same host. Fixes bug 637649.