dav: Unescape the path when setting the mount_prefix of the GMountSpec
As of commit 6e2f67 we escape the path of the mount_base. Later when we
set the mount_prefix of the backend's mount_spec we do use the mount_base
for this. We therefore have to unescape the path again because otherwise
comparing paths won't work in the client.