dav: Decode the paths of the multistatus response before comparing them
When doing a PROPFIND to stat () a location we have to find the "target"
response element that we did the PROPFIND for. We do this by comparing
the URL given by the href sub-element of the response element. Before
doing this comparison we have to decode the URL to make sure to ignore
encoding differences. Fixes bug 629660.