dav: Use default unmount logic
Get rid of our own unmount logic which was just calling exit () and
instead rely on the default unmount logic. The result is that first
we get busy-on-unmount automatically once bug 509606 is fixed;
also we are fixing bug 567664 since  by just calling exit () we never
sent a reply to the caller of unmount ().

https://bugzilla.gnome.org/show_bug.cgi?id=567664