ftp: Fix stupid proxy servers disconnecting on FEAT
A bunch of proxy servers assume that the first command sent by an FTP
server is USER and if it isn't, they bomb out. We do want to support
FEAT-before-USER though so we can do secure FTP (if we ever actually
implement it). So we need to make sure we try it at least once and if it
fails, we open a new connection and try with USER instead.