Fix symlink target in query_info_reply() for sftp backend
In SSH_FXP_NAME messages, the name field is preceded by a uint32
count field. query_info_reply() in gvfsbackendsftp.c did not take
the existence of the count field into account, and so set symlink
targets incorrectly.

For a simple demonstration, use gvfs-info to stat a symlink on an
ssh mount.

https://bugzilla.gnome.org/show_bug.cgi?id=652535