gvfs-open: Exit	with error code > 0 when open fails
Previously gvfs-open exited with success (error code 0) even when open
failed, either because file was not found or there was no default
application registred to open it. Without easy testing for error codes,
it was hard to use it in scripts. Now it exits with code 1 when any of
the given files are either not found or its default application is not
registered.

https://bugzilla.gnome.org/show_bug.cgi?id=655470