gdu: Don't ignore fstab mounts belonging to crypto volumes
Since we always show crypto volumes, we should never ignore
corresponding mounts even if they are supposed to be ignored.

If we consider the opposite case, i.e. ignoring volumes that have
corresponding mounts that should be hidden, we may end up with strange
behaviour like volume disappearing after mount/unlock. It's not
easily possible to determine that the particular unlocked volume
is supposed to be mounted in 'hidden' path.