Keep remote volume monitor proxies alive forever
Once constructed, it's a) more efficient; and b) simplifies the code;
if we keep the proxy for a remote volume monitor process around.

More importantly, this change sidesteps some thorny bugs in the
current implementation where the remote volume monitor proxy has been
disposed but not yet finalized - for some volume monitors, this can
happen not only pathological situations but also when running
e.g. gvfs-mount(1) to mount a volume.

https://bugzilla.gnome.org/show_bug.cgi?id=660295

Signed-off-by: David Zeuthen <davidz@redhat.com>