gphoto2: Initialize mutex early enough
Initializing the lock mutex in do_mount() is too late, as there are a lot of
code paths that call release_device() and thus g_mutex_clear() that run before
a successful do_mount(). For example, when you run gvfsd-gphoto2 with wrong
arguments, or gp_camera_init() fails in do_mount().

Now initialize it in g_vfs_backend_gphoto2_init().

https://bugzilla.gnome.org/show_bug.cgi?id=673893