udisks2: Set should_automount correctly when starting up
This bug caused volumes to not automount at session startup. While it
can be argued that we should or shouldn't at least this is

    1. a change from earlier versions of GNOME; and

    2. rather disruptive for some non-GUI users

Basically the problem is this: We need to take into account that on
coldplug, the heuristic of just looking at whether the volume has
appeared within the last five seconds won't work.

This bug report that triggered this bug fix was this one

 https://bugzilla.redhat.com/show_bug.cgi?id=813069

Signed-off-by: David Zeuthen <davidz@redhat.com>