Don't set should_automount to TRUE for devices on other seats
This change uses the libsystemd-login library to obtain the seat we're
on and compares it against the Seat property on the D-Bus interface
org.freedesktop.UDisks2.Drive. This property is available in udisks
1.95.0, see

 http://cgit.freedesktop.org/udisks/commit/?id=91106cdc7622d9674f6083dcb524407f026a36c7

Also, since we're still on the stable branch and not everyone may be
using systemd, make this work without hard-requiring either
libsystemd-login or udisks 1.95.0.

Signed-off-by: David Zeuthen <davidz@redhat.com>