udisks2: don't automount if drive is shared across all seats
If we automount media in a drive assigned to all seats, then all
sessions on all seats will be notified of the new media... which is
definitely going to be annoying. Instead, just disable automounting on
such drives so the user manually have to mount it.

Of course, this depends on the shell doing the right thing. Which is:
only notifying the user when successfully auto-mounting in response to
a hotplug event. This is not currently the case, see

 https://bugzilla.gnome.org/show_bug.cgi?id=660595#c19

for details about this bug.

Signed-off-by: David Zeuthen <davidz@redhat.com>