udisks2: Support getting/storing LUKS encryption passphrase from keyring
Also use a nicer message when asking for the passphrase, e.g. use strings like

 'SD04G (SD Card Reader)'
 'WD 2500JB External (250 GB Hard Disk)'

instead of /dev/mmcblk0p1 or /dev/sdb1. If stored in the keyring, we
also use the following display name

 'Encryption passphrase for $DRIVE'

where $DRIVE is of the above form. This makes it easy for the user to
manage (for example, delete) pass-phrases using seahorse(1).

This was discussed in bug 674161.

 https://bugzilla.gnome.org/show_bug.cgi?id=674161

This commit adds/changes translatable strings but Tomas said he would
soon branch gvfs for gnome-3-at a point before this patch.

Signed-off-by: David Zeuthen <davidz@redhat.com>