dav: abort finding the root on error or auth info change
The webdav root finding logic has been changed so it will not
continue to check the parent directory if we did not get a
webdav collection.
Also if we successfully obtained the information for a directory
we switch the auth handler to non-interactive so we don't try
to ask for new credentials for parent dir. Fix bug #677753.