fuse: Support ftruncate() where requested size > current size
When ftruncate is called with a size greater than the current size
of the file, the expected behavior is that the file size increases to
the requested size, and the new space is zero-filled.  SQLite depends
on this behavior when opening a database using 'write-ahead log'
journaling.

Based on original patch by Jeff Smith <whydoubt@yahoo.com>