Don't include jargon in mount display names and messages
The backend protocol isn't necessary in the display name for a
location. And having it there makes it much harder to actually
find the user interesting part of the name. Once a mount is connected
the user interesting part of the name is "what it is" and not
"how it is accessed".

This is possible now that we aren't using the display name
to make a unique mount point.

https://bugzilla.gnome.org/show_bug.cgi?id=682878