MTP: "Fix" pulling of directories.
No, this doesn't make recursive pull magically work, but it allows
a pull on a directory to make that directory, or silently succeed
if the directory exists already.

This does have the beneficial effect that file-roller can now
compress a directory from an mtp mount - if you ignore the fact
that it doesn't use display names.