gvfs-test: Drop usershares
We cannot rely on availability of usershares on all distros, so simplify the
tests to expect that gvfs-testbed sets up the "myfiles" share by itself and
then just use this.

As the same share now allows both guest and authenticated access, call
gvfs-mount with an "user@" prefix in the URI to tell it that we want
authenticated access, and adjust wait_for_gvfs_mount_user_prompt() to also
trigger on a Domain prompt (as gvfs-mount now does not ask for the user name
any more).