gvfs-test: Make Smb tests run without gvfs-testbed
When not running under gvfs-testbed, start smbd as normal user with temporary
configuration and state files on a high port (1445).

Set $LIBSMB_PROG to "netcat to port 1445" to convince libsmbclient and the smb
gvfs backend to talk to this port instead (unfortunately libsmbclient has no
API to set the port).

As in this mode smbd cannot change to other users it does not prevent access to
files which are only accessible to the test user, so adjust the test case
accordingly. This should be cleaned up later on by using two separate shares.

https://bugzilla.gnome.org/show_bug.cgi?id=686006