proxy volume monitor: Get session bus on demand
Do not connect to session bus at module load, let proxies to get
a bus when needed. No need to keep it active all the time. Proxies
will return an error when a bus is unavailable.

Also don't close the shared d-bus connection on module unload, let
it be managed by proxies.

https://bugzilla.gnome.org/show_bug.cgi?id=687074