gvfs-test: Fix Drive tests for Python 3.3
Python 3.3 changed the behaviour of subprocess.Popen() with
universal_newlines=True: now stdin is expected to be a str as well, which makes
sense, but didn't work that way in Python 3.2.

Change the code to explicitly recode and not use universal_newlines=True, so
that this works for both Python versions.