gvfs-test: Skip Sftp.test_unknown_host when not in gvfs-testbed
When running under the normal user account, the StrictHostKeyChecking option is
client-configurable behaviour which cannot be temporarily changed. Skip the
test entirely in this case instead of exfailing it, as it does not test
anything meaningful then.

Drop the exfail so that the test must succeed when running under gvfs-testbed.