MTP: Review Feedback: Correctly handle push/pull requests for directories.
Tomas explained the correct way to tell Nautilus to recurse directories
and issue individual copy operations for the contents, and it works!

Note that in the case of push, the contents don't copy correctly as
the push path is incorrect (uses filename instead of file id). I
can compensate for this, but that will be a separate change.