MTP: Fix segfault if client queries for non-existent file by name.
If the client, usually Nautilus, ends up trying to find a file
by name instead of ID, there might not be a match. We should
return NOT_FOUND explicitly, as there is no libmtp error generated
in this case.

This situation arises after:
* A file is renamed
* A directory is created
* A file is created (which actually fails)

Refreshing the Nautilus view will get Nautilus back in sync.