daemon: Implement proper org.gtk.vfs.MountTracker.UnregisterMount()
A proper unmount method was missing from the daemon interface as it was
basically not needed. The mount tracker was watching mounts on the bus
and if child disappeared, that event was treated as unmount.

With recent introduction of gvfs test suite we started using
man-in-the-middle program for connecting to private SMB server. However,
certain libsmbclient versions are forking gvfs backends (gvfsd-smb in
this case) which makes d-bus confused and name owner tracking doesn't
work as expected.

Moreover the forked process got stuck due to internal samba LIBSMB_PROG
handling after the original process did its job and unmount was triggered
from user side. The original process exited just fine but the forked one
was blocking the name on the bus probably and gvfs mount tracker was fooled.
Eventually the stuck libsmbclient call times out and exits gracefully.
This happens with samba-3.6.9

This patch adds proper unmount call to indicate the backend is going away
so even if there are leftovers, user is indicated a proper unmount.

https://bugzilla.gnome.org/show_bug.cgi?id=691568