gvfs-test: Don't wait for GVolumeMonitor "mount-removed" signal
With implementing proper unmount method in the last commit the mount
should be gone when GVfsJobUnmount returns. It's not guaranteed when
the "mount-removed" signal is received to the application.

Let's not wait for the "mount-removed" signal, a return value from
g_file_unmount_mountable_with_operation_finish() should be enough to
indicate success or error.

https://bugzilla.gnome.org/show_bug.cgi?id=691568