channel: Unqueue cancelled requests
We put a channel request on the output buffer and start writing, but
if the write is cancelled on the first call (i.e. no partial writes)
we abort immediately without ever writing the request.

However, if we do this we also need to unqueue the request from the output
buffer, as otherwise this will be sent with the next operation. This
can be problematic for seeks as the seek generation is then not in sync.

https://bugzilla.gnome.org/show_bug.cgi?id=675181