gvfs-test: Add first test for gphoto2 backend
Add umockdev dump and ioctl trace for a Canon PowerShot SX200 digicam, with
ioctls from these commands recorded:

  gvfs-mount -li
  gvfs-mount gphoto2://[usb:001,011]/
  gvfs-info gphoto2://[usb:001,011]
  gvfs-info gphoto2://[usb:001,011]/DCIM
  gvfs-info gphoto2://[usb:001,011]/DCIM/100CANON/IMG_0001.JPG
  gvfs-info gphoto2://[usb:001,011]/DCIM/100CANON/IMG_0002.JPG
  gvfs-cat gphoto2://[usb:001,011]/DCIM/100CANON/IMG_0001.JPG
  gvfs-cat gphoto2://[usb:001,011]/DCIM/100CANON/IMG_0002.JPG

(With two tiny test images, not actual big photos)

If umockdev is available, use it to simulate that device and check that we can
get the directory and file info, and access the pictures.

Note that this does not yet work under gvfs-testbed, as we somehow need to
inject the $UMOCKDEV_ROOT into the spawned d-bus.