gvfs-test: Add first test for mtp backend
Add umockdev dump and ioctl trace for a Sony Xperia Mini mobile phone (with
CyanogenMod), with some small demo files and the ioctls from these commands
recorded:

  gvfs-mount 'mtp://[usb:001,017]'
  gvfs-ls 'mtp://[usb:001,017]/SD-Karte'
  gvfs-ls 'mtp://[usb:001,017]/SD-Karte/Music'
  gvfs-ls 'mtp://[usb:001,017]/SD-Karte/Music/GStreamer - The Test Sine'
  gvfs-info 'mtp://[usb:001,017]/SD-Karte/Music/GStreamer - The Test Sine'
  gvfs-info 'mtp://[usb:001,017]/SD-Karte/Music/GStreamer - The Test Sine/sine.ogg'
  gvfs-cat 'mtp://[usb:001,017]/SD-Karte/Music/GStreamer - The Test Sine/sine.ogg'
  gvfs-cat 'mtp://[usb:001,017]/SD-Karte/hello.txt'
  gvfs-info 'mtp://[usb:001,017]/SD-Karte/hello.txt'
  gvfs-info 'mtp://[usb:001,017]/SD-Karte/DCIM/100CANON/IMG_0001.JPG'
  gvfs-cat 'mtp://[usb:001,017]/SD-Karte/DCIM/100CANON/IMG_0001.JPG'

If umockdev is available, use it to simulate that device and check that we can
get the directory and file info, and access the file contents.

This bumps the (optional) umockdev dependency to >= 0.2.10.