daemon: Emit signal before returning dbus value
In gvfsjobopenforread.c, the dbus method returns a value in create_reply
which eventually results in a GVfsJobRead job to be sent to the backend.
This could happen before the "new-source" signal is emitted.  If this
happens, the job is not queued because the "new_job" signal would not
have been connected to a job source yet.  The read then hangs because
the GVfsJobRead is lost.

This is hit often when performing large smb transfers (see
https://bugzilla.gnome.org/show_bug.cgi?id=697782 and
https://bugs.launchpad.net/ubuntu/+source/gvfs/+bug/1075923).
It can be reproduced by putting a small sleep before the
g_signal_emit_by_name call.

Fix this by emitting the "new-source" signal *before* the dbus method
value is returned.  This ensures that the "new_job" signal is set up
before any further jobs are sent.

Note that the same problem and solution applies for
gvfsjobopenforwrite.c.