sftp: Implement try_query_fs_info using the OpenSSH statvfs extension
Use the statvfs command from OpenSSH to fill in the filesystem size,
free and used attributes as well as whether the filesystem is readonly
or not.

The extension is defined at http://api.libssh.org/rfc/PROTOCOL.
The extension is signified by the name "statvfs@openssh.com" with a
version of "2".

Based on a patch Alban Browaeys.

https://bugzilla.gnome.org/show_bug.cgi?id=549207