daemon: Make progress callbacks work with try methods
Before, for each job type that implements progress callbacks, the
progress_proxy was being unref'd immediately after the backend's try or
run method had run.  This means that if the backend had an asynchronous
callback which updated the progress, it wouldn't work because the
progress_proxy had been removed.

Fix this by unrefing the progress_proxy in the GVfsJobProgress class's
finalize method.

https://bugzilla.gnome.org/show_bug.cgi?id=711247