MTP: Fail fast if in the middle of an unmount
I've seen a ton of bug reports where the backend crashes due to
operations executing after an unmount begins. I think it's a
sufficient solution to check the unmount flag that we already have
and then immediately abort the operation.

Generally, this is only seen with operations that are initiated
implicitly like do_query_info or do_enumerate, but I've added the
protection to all operations for consistency.