dav: Use the native MOVE operation for moving
This saves unnecessary round trips when moving files and directories.
Based on a patch by Mads Chr. Olesen and the GLocalFile implementation.

Backups are not yet implemented.  It tries to cover the various cases of
overwriting files with directories and vice versa by doing what
GLocalFile does.  Notably, webdav servers are capable of overwriting a
directory with a file which is not a supported operation for
g_file_move, while at least Apache's mod_dav gives a generic 400 Bad
Request when trying to overwrite a file with a directory.

https://bugzilla.gnome.org/show_bug.cgi?id=572786